+++ tests/functional/registration/test_version.py
@@ -7,11 +7,11 @@ from packaging.version import parse
 from textx.cli import textx
 
 
-def test_version_command():
+def test_version_command(caplog):
     runner = CliRunner()
     result = runner.invoke(textx, ["version"])
     assert result.exit_code == 0
-    assert result.output.startswith("textX")
-    version_text = result.output.split()[-1]
+    assert 'textX' in caplog.text
+    version_text = caplog.text.split()[-1]
     version = parse(version_text)
     assert version.__class__.__name__ == "Version"